[CD] [69842562] Expect "callable" instead of class
In order to be more flexible with our API, expect the configuration
service adapter to respond to `call` instead of `new` and `get_config`.
This allows us to provide anything that is "callable", e.g., a block,
proc, lambda, a class with a `call` method, etc.

Also, the `use_config_service` boolean isn't required; remove it.